Adds wrappers for proving
